+++ src/track.h
@@ -923,14 +923,14 @@ bool tracker_channel_disable(tracker_channel_id_t id);
 /* Tracking parameters interface. */
 
 void tracking_channel_measurement_get(
-    double ref_tc,
+    u64 ref_tc,
     const tracking_channel_info_t *info,
     const tracking_channel_freq_info_t *freq_info,
     const tracking_channel_time_info_t *time_info,
     const tracking_channel_misc_info_t *misc_info,
     channel_measurement_t *meas);
 
-bool tracking_channel_calc_pseudorange(double ref_tc,
+bool tracking_channel_calc_pseudorange(u64 ref_tc,
                                        const channel_measurement_t *meas,
                                        double *raw_pseudorange);
